Move get_call_data syscall into package
